+++ ModManager/Gui/Main.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -78,17 +79,37 @@ namespace ModManager.Gui
             ListViewItem listItem;
 
             ModViewModel modMeta = mod.Value;
-            listItem = new ListViewItem(new string[] 
+            if (modMeta != null)
             {
-                modMeta.Caption,
-                modMeta.Downloaded,
-                modMeta.SupportedVersions,
-            });
+                listItem = new ListViewItem(new string[] 
+                {
+                    modMeta.Caption,
+                    modMeta.Downloaded,
+                    modMeta.SupportedVersions,
+                });
+            }
+            else
+            {
+                listItem = new ListViewItem(new string[]
+                {
+                    mod.Key,
+                    null,
+                    null,
+                });
+            }
 
 
             listItem.Name = mod.Key;
-            if (listItem.SubItems[2].Text.Contains(_presenter.ActiveMods["ludeon.rimworld"].SupportedVersions) == false)
+            if (modMeta == null)
+            {
+                listItem.ToolTipText = "This mod is missing";
                 listItem.BackColor = _presenter.IncompatibleColor;
+            }
+            else if (listItem.SubItems[2].Text.Contains(_presenter.ActiveMods["ludeon.rimworld"].SupportedVersions) == false)
+            {
+                listItem.ToolTipText = "Incompatible with current version.";
+                listItem.BackColor = _presenter.IncompatibleColor;
+            }
 
             return listItem;
         }
@@ -150,6 +171,8 @@ namespace ModManager.Gui
 
             menuStrip1.Enabled = false;
             _presenter.SelectedMod = null;
+            DescriptionWebBrowser.DocumentText = String.Empty;
+            PresenterBindingSource.ResetBindings(false);
         }
 
         private void RefreshButton_Click(object sender, EventArgs e)
@@ -211,17 +234,14 @@ namespace ModManager.Gui
             foreach (var item in ActiveModsListView.Items.Cast<ListViewItem>())
             {
                 if (item.BackColor == _presenter.IncompatibleColor)
-                {
-                    item.ToolTipText = "Incompatible with current version.";
-                    continue;
-                }
+                    continue; 
+
 
                 ModViewModel mod;
                 _presenter.ActiveMods.TryGetValue(item.Name, out mod);
 
                 if (mod == null)
                     _presenter.AvailableMods.TryGetValue(item.Name, out mod);
-
 
                 StringBuilder tooltip = new StringBuilder();
 
@@ -263,9 +283,6 @@ namespace ModManager.Gui
                 else
                     item.BackColor = Color.Transparent;
             }
-
-
-
         }
 
         private void ModsListView_ColumnClick(object sender, ColumnClickEventArgs e)
@@ -354,6 +371,37 @@ namespace ModManager.Gui
                 System.Diagnostics.Process.Start(path);
             else
                 MessageBox.Show("Browse folder", "Folder does not exist.", MessageBoxButtons.OK);
+        }
+
+        private void exportToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            using (SaveFileDialog dialog = new SaveFileDialog())
+            {
+                dialog.AddExtension = true;
+                dialog.Filter = "mod list (*.modlist)|*.modlist";
+
+                if (dialog.ShowDialog(this) == DialogResult.OK)
+                    _presenter.ExportModlist(dialog.FileName);
+            }
+        }
+
+        private void importToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            string savePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "AppData", "LocalLow", "Ludeon Studios", "RimWorld by Ludeon Studios", "Saves");
+            using (OpenFileDialog dialog = new OpenFileDialog())
+            {
+                dialog.AddExtension = true;
+                dialog.Filter = "All|*.modlist;*.rws|mod list (*.modlist)|*.modlist|rimworld save (*.rws)|*.rws";
+                dialog.CheckFileExists = true;
+                dialog.Multiselect = false;
+                dialog.FilterIndex = 1;
+
+                if (Directory.Exists(savePath))
+                    dialog.InitialDirectory = savePath;
+
+                if (dialog.ShowDialog(this) == DialogResult.OK)
+                    _presenter.ImportModlist(dialog.FileName);
+            }
         }
     }
 }