+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-Added option for rolling backups
-Fixed error message when activating mods.
-Fixed incorrent numbers on incompatible mods warning message.
+Improved support for mod relationships
+Improved handling of missing mods
+Added Import/Export functionality:
+- Import list of active mods from either a rimworld save or an exported mod list
+- Export active mods as an importable mod list file